requirements.txt
asgiref==3.8.1
asttokens==3.0.0
decorator==5.1.1
Django==4.2.18
django-allauth==65.3.1
exceptiongroup==1.2.2
executing==2.2.0
ipython==8.18.1
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
